Implement efficient quiz loading (see description)
This update changes the way quizzes are loaded: they are downloaded, then saved to local storage, where they can later be retrieved across the site.

If a quiz is updated, the site detects that (due to the fact that the "created" metadata changes after an edit is made), and downloads the updated quiz.

This reduces downloads signifigantly.